+++ apps/website/src/components/Header.tsx
@@ -64,9 +64,9 @@ export default function Header() {
 	);
 
 	return (
-		<header className="dark:bg-dark-400 dark:border-dark-100 bg-light-600 border-light-800 fixed top-0 left-0 z-20 w-full border-b">
+		<header className="dark:bg-dark-600/75 dark:border-dark-100 border-light-900 sticky top-4 z-20 rounded-md border bg-white/75 shadow backdrop-blur-md">
 			<div className="block h-16 px-6">
-				<div className="flex h-full flex-row place-content-between place-items-center">
+				<div className="flex h-full flex-row place-content-between place-items-center gap-8">
 					<Button
 						aria-label="Menu"
 						className="focus:ring-width-2 focus:ring-blurple flex h-6 w-6 transform-gpu cursor-pointer select-none appearance-none flex-row place-items-center rounded border-0 bg-transparent p-0 text-sm font-semibold leading-none no-underline outline-0 focus:ring active:translate-y-px lg:hidden"
@@ -74,21 +74,21 @@ export default function Header() {
 					>
 						<VscMenu size={24} />
 					</Button>
-					<div className="hidden md:flex md:flex-row md:overflow-hidden">{breadcrumbs}</div>
-					<div className="flex flex-row place-items-center gap-4">
-						<Button
-							as="div"
-							className="dark:bg-dark-800 focus:ring-width-2 focus:ring-blurple rounded bg-white px-4 py-2.5 outline-0 focus:ring"
-							onClick={() => dialog?.toggle()}
-						>
-							<div className="flex flex-row place-items-center gap-4">
-								<VscSearch size={18} />
-								<span className="opacity-65">Search...</span>
-								<div className="md:opacity-65 hidden md:flex md:flex-row md:place-items-center md:gap-2">
-									<FiCommand size={18} /> K
-								</div>
+					<div className="hidden lg:flex lg:flex-row lg:overflow-hidden">{breadcrumbs}</div>
+					<Button
+						as="div"
+						className="dark:bg-dark-800 focus:ring-width-2 focus:ring-blurple w-56 grow rounded bg-white px-4 py-2.5 outline-0 focus:ring sm:grow-0"
+						onClick={() => dialog?.toggle()}
+					>
+						<div className="flex flex-row place-items-center gap-4 md:justify-between">
+							<VscSearch size={18} />
+							<span className="opacity-65">Search...</span>
+							<div className="md:opacity-65 hidden md:flex md:flex-row md:place-items-center md:gap-2">
+								<FiCommand size={18} /> K
 							</div>
-						</Button>
+						</div>
+					</Button>
+					<div className="flex flex-row place-items-center gap-4">
 						<Button
 							aria-label="GitHub"
 							as="a"