+++ apps/website/src/components/Property.tsx
@@ -25,10 +25,9 @@ export function Property({
 
 	return (
 		<div className="scroll-mt-30 flex flex-col gap-4" id={item.displayName}>
-			<div className="md:-ml-8.5 flex flex-col gap-2 md:flex-row md:place-items-center">
-				<Anchor href={`#${item.displayName}`} />
+			<div className="flex flex-col gap-2 md:-ml-9">
 				{isDeprecated || item.isReadonly || item.isOptional ? (
-					<div className="flex flex-row gap-1">
+					<div className="flex flex-row gap-1 md:ml-7">
 						{isDeprecated ? (
 							<div className="flex h-5 flex-row place-content-center place-items-center rounded-full bg-red-500 px-3 text-center text-xs font-semibold uppercase text-white">
 								Deprecated
@@ -47,14 +46,19 @@ export function Property({
 					</div>
 				) : null}
 				<div className="flex flex-row flex-wrap place-items-center gap-1">
+					<Anchor href={`#${item.displayName}`} />
 					<h4 className="break-all font-mono text-lg font-bold">
 						{item.displayName}
 						{item.isOptional ? '?' : ''}
 					</h4>
-					<h4 className="font-mono text-lg font-bold">{separator}</h4>
-					<h4 className="break-all font-mono text-lg font-bold">
-						<ExcerptText excerpt={item.propertyTypeExcerpt} model={item.getAssociatedModel()!} />
-					</h4>
+					{item.propertyTypeExcerpt.text ? (
+						<>
+							<h4 className="font-mono text-lg font-bold">{separator}</h4>
+							<h4 className="break-all font-mono text-lg font-bold">
+								<ExcerptText excerpt={item.propertyTypeExcerpt} model={item.getAssociatedModel()!} />
+							</h4>
+						</>
+					) : null}
 				</div>
 			</div>
 			{hasSummary || inheritedFrom ? (