+++ apps/website/src/components/model/enum/Enum.tsx
@@ -4,14 +4,12 @@ import { Panel } from '../../Panel';
 import { Documentation } from '../../documentation/Documentation';
 import { ObjectHeader } from '../../documentation/ObjectHeader';
 import { DocumentationSection } from '../../documentation/section/DocumentationSection';
-import { SummarySection } from '../../documentation/section/SummarySection';
 import { EnumMember } from './EnumMember';
 
 export function Enum({ item }: { item: ApiEnum }) {
 	return (
 		<Documentation>
 			<ObjectHeader item={item} />
-			<SummarySection item={item} />
 			<DocumentationSection icon={<VscSymbolEnum size={20} />} padded title="Members">
 				<div className="flex flex-col gap-4">
 					{item.members.map((member, idx) => (