+++ apps/website/src/util/constants.ts
@@ -4,6 +4,7 @@ export const PACKAGES = [
 	'collection',
 	'core',
 	'formatters',
+	'next',
 	'proxy',
 	'rest',
 	'util',
@@ -15,8 +16,10 @@ export const N_RECENT_VERSIONS = 2;
 
 export const OVERLOAD_SEPARATOR = ':';
 
+export const METHOD_SEPARATOR = '#';
+
 export const DESCRIPTION =
-	"discord.js is a powerful node.js module that allows you to interact with the Discord API very easily. It takes a much more object-oriented approach than most other JS Discord libraries, making your bot's code significantly tidier and easier to comprehend.";
+	"discord.js is a powerful Node.js module that allows you to interact with the Discord API very easily. It takes a much more object-oriented approach than most other JS Discord libraries, making your bot's code significantly tidier and easier to comprehend.";
 
 export const CODE_EXAMPLE = `import { Client, GatewayIntentBits } from 'discord.js';