+++ packages/discord.js/src/structures/UserSelectMenuBuilder.js
@@ -13,15 +13,12 @@ class UserSelectMenuBuilder extends BuildersUserSelectMenu {
   }
 
   /**
-   * Creates a new select menu builder from json data
-   * @param {JSONEncodable<APISelectMenuComponent> | APISelectMenuComponent} other The other data
+   * Creates a new select menu builder from JSON data
+   * @param {UserSelectMenuBuilder|UserSelectMenuComponent|APIUserSelectComponent} other The other data
    * @returns {UserSelectMenuBuilder}
    */
   static from(other) {
-    if (isJSONEncodable(other)) {
-      return new this(other.toJSON());
-    }
-    return new this(other);
+    return new this(isJSONEncodable(other) ? other.toJSON() : other);
   }
 }
 
@@ -29,5 +26,5 @@ module.exports = UserSelectMenuBuilder;
 
 /**
  * @external BuildersUserSelectMenu
- * @see {@link https://discord.js.org/#/docs/builders/main/class/UserSelectMenuBuilder}
+ * @see {@link https://discord.js.org/docs/packages/rest/stable/UserSelectMenuBuilder:Class}
  */